azure-func/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
aiohttp
requests
pytz
celery==5.4.0
uvicorn
redis
aioredis
pytz
python-dotenv